requirements.txt
pandas
azure-functions
azure-functions-worker
matplotlib